[Fixes #12076] Move the new users API under API module
